[SYCL] Add a partial profiling workaround for acc devices (#12359)
With the recent fixes to command submit profiling info, queue profiling is no longer supported on devices with OpenCL version < 2.1. This primarily affects applications on accelerator devices. The full solution would be to provide a fallback implementation of profiling information that is less precise but does not require querying device time. As a temporary workaround, this patch allows to create an accelerator queue with profiling despite the lack of queue_profiling aspect and instead throw an error if submit profiling info is requested.