Update Fortran RTL call for initializing private objects (#10625)
* Update Fortran RLT call for initializing private objects

Fortran objects with a dope vector need the dope vector initialized,
and code generated by ParOpt assumes that the init routine returns
the object's size.  When the object is unallocated, the IL0 compiler
returned 0.  There are certain cases where 0 is a valid size for an
allocated Fortran array, so the init routine has been updated to
return -1 for unallocated instead of 0.

This patch calls the updated routine, `_f90_dope_vector_init2` and
expects a return value of -1 when the object is not allocated.

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Return 0 when _f90_dope_vector_size says array is unallocated

Fortran allows zero length unallocated arrays, but ParOpt does not
make this distinction, so return zero for the size of an unallocated
array, as well as for zero length arrays (at ParOpt's request).

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Privatize zero length arrays for Fortran

Fortran allows zero length arrays, so alloca space for them even
if it is zero bytes.

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Formatting changes requested by clang-format

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* For backwards compatibility store 0 in NumElements when not allocated

The Item passed to getF90DVInitCode stores and sometimes uses NumElements.
Make sure it is always defined and set to zero for unallocated for
backward compatibility.  Tests with arrays allocated with length 0 seem
to work, so probably it is OK to use NumElements = 0 for both unallocated
and zero-length arrays.

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Fix indenting

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Changes for code review comments

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Make LIT tests expect new Fortran DV init function

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Prevent unnecessary guard barrier for _f90_dope_vector_init2

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Reflow list to make clang-format happy

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Retain old behavior for zero-length arrays

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

* Make LIT tests match change from SGE to SGT

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

---------

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>