[Driver][SYCL] Update option for undefined symbols for hostdep link (#8947) (#12421)
When performing the host link for dependencies, we added '-z undefs' to
the link command line. This is fine for the bfd linker, but the option
is not supported for gold. Update the option used to use the common
--unresolved-symbols=ignore-all instead.

Addresses: CMPLRLLVM-46474
Cherry-pick of: https://github.com/intel/llvm/commit/554ed2187d320bb3d18616a439e9b867a5aff10d