Update failed LIT test (#12495)
Update test for memory fold tables after #10680
// INTEL_FEATURE_ISA_APX_F
NOTE: Updates of the APX fold pairs are wrong in this PR due to incorrect definitions in TD introduced by #10680. The correct fix is not straightforward, so I provide a hot fix (workaround) here to let it not block other XISA PRs. Since the incorrect pairs are all about APX, it does not affect non-APX developers.
// end INTEL_FEATURE_ISA_APX_F
JIRA: CMPLRLLVM-46899