[SYCL] Set compiled program for hostpipe (#12631)
* [SYCL] Set compiled program for hostpipe

HostPipe implementation should pass the compiled program to backend
when invoke the read / write operation.

There are two scenarios for HostPipe read / write:
  * Invoke device kernel before HostPipe read / write
  * Invoke HostPipe read / write before device kernel.
For the 1st scenario, the program should be compiled before HostPipe
read / write. We get the program from Cache for HostPipe entry.

For the 2nd scenario, the program should NOT be compiled. We need to
compile the program first for HostPipe entry. The compiled program will
be cached for future device kernel.

We also added the test cases for this feature.

* format test cases