[VPlan][CallVecDecisions] Fix to get stride for truncated value (#12564)
This patch fixes an issue where the value stored to private memory
for uval/val arguments is truncated. What happens is that DA says
that the truncated value is random shape, which leads to a caller/
callee mismatch because the variable stride caller side variants
are only created. We need to match the constant stride callee variant.
In order to generate the constant stride caller variants, the shape
of the truncated operand is used instead. The same change was made
previously for truncated integer args and the same utility is used
here.