[LICM] For related load speculation, use number of loads with TBAA to limit its cost. (#12449)
TBAA info is used as one of the criteria for related load speculation. But the
TBAA metadata of the hoisted loads under conditionals are dropped. This can
inhibit further speculated hoisting of loads from the same struct because we
can't find a eligible related load within the cost limit after a certain
number of instructions are hoisted.

This patch modifies cost calculation to count only loads with TBAA metadata
attached, and reduces the value of maximum cost correspondingly.

JIRA: CMPLRLLVM-46423