[SYCL] Fix integration footer for specialization_id  (#8959) (#12424) (#12633)
Type printer for `specialization_id` in SemaSYCL did not
correctly handle class template specializations. The
template parameters were omitted in generated shim
functions i.e. for the following code,
```
namespace {
  namespace inner {
    template <int Val>
    struct S3 {
      static constexpr sycl::specialization_id a{Val};
    };
  }
}
template class inner::S3<10>;
```

`inner::S3::a` was generated in shim functions instead of
`inner::S3<10>::a`. This patch fixes this error

This is a cherry-pick of #12424
Fixes: CMPLRLLVM-45480