[NFC] [DTrans] Fix root cause of memory leak in DTransSafetyAnalyzer (#12567)
* [NFC] [DTrans] Fix root cause of memory leak in DTransSafetyAnalyzer

This reworks the fix in https://github.com/intel-restricted/applications.compilers.llvm-project/pull/11881 to fix the root cause of TypeInfo objects
getting created that were not destroyed by preventing the extra objects from
being created in the first place. The problem was that when a TypeInfo object
is being created, it can trigger the creation of additional TypeInfo objects,
such as for types stored within a structure, and the construction of these
types could lead back to the creation of the object that triggered their
creation, so before creating the object, a check is added to ensure the type
was not already created. Otherwise, this could lead to two TypeInfo objects
existing for the same Type, but some objects could refer to one instance,
while other objects referred to the other instance.

Also fixed one assertion string typo discovered while investigating.

[CMPLRLLVM-45023]

* Fixes for Robert

---------

Co-authored-by: christopher.m.chrulski <cmchruls@scsel-clx-13.sc.intel.com>