[Driver][SYCL] Update option for undefined symbols for hostdep link (#8947) (#12421) (#12637)
When performing the host link for dependencies, we added '-z undefs' to
the link command line. This is fine for the bfd linker, but the option
is not supported for gold. Update the option used to use the common
--unresolved-symbols=ignore-all instead.

Addresses: CMPLRLLVM-46474
Cherry-pick of: 45f6a1cb4b0772d57a93b0b78cf209e2dfe47910