[SYCLTransforms] Check whether one kernel is called from NDRange (#12797)
The max_global_work_dim with value 0 is set to single_task kernels
for SPIR FPGA target. But the logic to check whether one kernel is
called from NDRange by max_global_work_dim value in PipeOrdering
pass is not correct. The kernel from single_task is not determined
to be called from NDRange.

As all kernels from SYCL language are called from NDrange. We directly
check the source in PipeOrdering pass.