[InferAttrs] Mark frexp and modf as memory(argmem: write) (#12689)
These two math library functions can't write to errno, but they do
produce two results and store the second result to a pointer passed as
their second argument. Appropriately marking them as
memory(argmem: write) enables more optimizations on calls to them.

Reviewed By: xbolva00

Differential Revision: https://reviews.llvm.org/D148043

This change was merged upstream as 750e20e. I did a manual cherry-pick
to xmain in #12514, but the commit I merged there (37e0181) was empty
because the pulldown is fast these days and merged the original change
to xmain first (with a conflict resolution in #12518).

This change is needed for 2023.2 as part of the fix for CMPLRLLVM-46501.