[VPlan] Fix bug with live-out adjustments for multiple-live-out inductions (#12686)
When an induction needs a closed form created, we create a new induction
operation to replace the original one, and fix up live-outs for the
original induction operation.  However, when later looking for the exit
instruction for the induction, we still look for the original induction
operation, which is no longer live-out.  When the induction has another
linked instruction that is live-out, we incorrectly treat it as the exit
instruction.  Fix that by updating the stored induction operation.