[NFCI] Extract getNameAndDbgLoc() into a method on DDRef. (#12860)
In `DDGraph.cpp` there is a function getNameAndDbgLoc() that is used to
print the name and debug location of the DDRef for printing assumed
dependences in the opt report. This patch extracts that function into a
method on DDRef named getNameAndDbgLocForOptRpt() so that it may be used
in other contexts.