[NFC][VPlan] Tunable cost heuristic for unrolling (#12455)
Updates the unrolling cost model heuristic to adjust the cost reduction for partial sums based on an estimated ILP benefit.

The getReductionCost implementation now collects the costs for the reduction chain (reachable by depth-first traversal of PHI uses) and other operand costs, as well as a coefficient (presently just set by an option) that can be set for each partial sum candidate while computing these costs.

The unroll heuristic first estimates the ILP benefit based on the ratio of the total reduction cost vs. the incoming loop cost, which is then multiplied by the precomputed coefficient. This value is then used to adjust the reduction chain cost - not the total cost of all operands - so that after multiplying by UF, the cost for the reduction chain is between 1 to UF the TTI costs.

This also adds an overhead cost estimate for the post-loop reduction for partial sum candidates. The cost adjustment for the latch branch is unchanged.