[VPlan][OptRpt] Add peeling memref to debug opt report. (#12794)
When dynamically peeling, we select a memref to dynamically align when
emitting the peel loop, e.g:

```
   #pragma vector dynamic_align
   for (unsigned I = 1; I < N; ++I)
     B[I] = A[I - 1] + A[I] + ...;
     ^
      \---- will be aligned by peel loop
```

This patch adds a debug remark to the opt-report about that memref's
debug location (line:column), as well as whether or not the memref we
peeled by was a load or a store (to assist if the debug location is
imprecise). This remark is added into the peeled loop's section, e.g:

```
LOOP BEGIN at test.c (11, 4)
<Peeled loop for vectorization>
    remark #15576: peel loop is dynamic
    remark #15577: estimated number of scalar loop iterations peeled: 3
    remark #15578: DEBUG: peeled by memref at 12:34 (store)
LOOP END
```

This serves as a first implementation, and is intentionally not exposed
to the user. The intent is to eventually improve debug locations so that
this information can be exposed to the user (similar to how it is shown
in mainline.)

JR: CMPLRLLVM-46335