[VPlan][CM][Compile Time][NFC] Reduce LiveValues map size. (#12898)
SpillFill heuristic maintains LiveValues DenseMap which maps the register
pressure a VPIntruction imposes on VPInstruction. When the values produced
by VPInstruction dies we delete an element from this map rather than set
it to 0. It reduces the map and boosts loops iterating through all elements
of the map.

Fixes Jira: CMPLRLLVM-44241
xmain PR: #12758