[VPlan] Handle variadic arguments for bailout remarks (driver portion) (#12944)
The current bailout remarks design doesn't handle multiple arguments to
the remark format string particularly well.  This is already proving
fragile.  This patch redesigns the bailout data to be stored within an
OptRemark object, which already handles variadic data for strings and
integers.

This patch only updates the VPlan driver code to use the new methodology.
Once it's accepted, the related changes to legality, planner, and code
generation can be developed, and we can delete support for the previous
design.  Code that will be deleted is marked as TODO in this patch.

Reference: CMPLRLLVM-46887