[DTrans] Extend DTrans dead metadata removal pass to handle function signature attributes (#13116)
* [DTrans] Extend DTrans dead metadata removal pass to handle function signature attributes

This is basically a relanding of https://github.com/intel-restricted/applications.compilers.llvm-project/pull/12864
now that the https://github.com/intel-restricted/applications.compilers.llvm-project/pull/13080 has landed
which was needed to avoid a performance regression being triggered by this PR.

Extend the metadata removal pass to clean up elements in the function signature
metadata to account for unreferenced values. For example:
define i64 @test1(ptr "intel_dtrans_func_index" = "2" % in,
ptr "intel_dtrans_func_index" = "4" % in2)
!intel.dtrans.func.type !7 {
where
!7 = distinct !{!3, !4, !5, !3}

In this case, the metadata node refers to types that have been deleted from
the function signature. (It is by design that the locations that remove
function parameters do not need to update the metadata as a way of avoiding
a lot of locations in the compiler from needing to know about the specific
encoding of DTransTypeMetadata). This function will rewrite the metadata to
be:
define i64 @test1(ptr "intel_dtrans_func_index"="1" %in,
ptr "intel_dtrans_func_index"="2" %in2)
!intel.dtrans.func.type !2 {
where
!2 = distinct !{!3, !4}

This is needed in some cases to avoid structure types from only being present
in the metadata, and not reachable from elsewhere in the IR.

[CMPLRLLVM-47260]

* Change comment grammar for Robert