Allow formation of SCCs whose root node has range info (#12922)
Previously, we bailed out on SCCs whose root node has range info because this
root node is substituted in other SCC insts in HIR. The substitution can result
in incorrect simplification by ScalarEvolution using the range info. There is
an example of this incorrect substitution in the code comments.

We fix this issue by creating a new root node during SSA deconstruction which
will be used for substitution.