[Paropt][Offload] Fix loop collapsing for generic loop (#13140)
JIRA: CMPLRLLVM-41323

For "target teams distribute parallel for collapse(2)" we don't
collapse the loop so we can optimize it with NDRANGE of 2 dim.
The same should be done to "target teams loop collapse(2)"
when "loop" is replaced with "distribute parallel for".
However, this was not happening because collapsing is done
before loop replacement. The fix was for collapsing to
look ahead at the loop replacement decision so it has the
info needed to decide whether to collapse or not.