[SYCL] Correctly handle special values in fast math mode (#9278) (#13174)
This PR fixes
[stream](https://github.com/intel/llvm/blob/sycl/sycl/test-e2e/Basic/stream/stream.cpp#L1)
tests that resulted in hangs for CUDA.

An implication of using `-ffast-math` is that the compiler is not
required to honour `nan` and `inf`. As a result calls to
[checkForInfNan](https://github.com/intel/llvm/blob/sycl/sycl/include/sycl/stream.hpp#L378)
were optimized away breaking the logic of `ScalarToStr`. The patch
provides a manual way of detecting `nan` and `inf` when in fast math
mode.

---
Cherry pick for codeplay for 2023.2
Co-authored-by: Jakub Chlanda <jakub@codeplay.com>