[NFCI][VPlan][HIR] Add an interface to set GEP ref specifics and start using the same (#13203)
During HIR vector code generation when creating GEP refs, we need to ensure GEP ref
specifics such as symbase are set appropriately. This is done currently by preserving
symbases during VPlan construction and using the same during HIR vector code generation.
The generated refs also need to have the number of collapsed levels set correctly
for correctness(CMPLRLLVM-45451). This PR is an NFC change to start using a new
interface to set GEP specifics which can be extended to also handle other specifics
such as number of collapsed levels. A follow on PR stores/sets the number of collapsed
levels to fix part of the issue reported in CMPLRLLVM-45451.