[VPlan][CM] Enable separate threhold for SW Gathers/Scatters. (#13123)
With the change software emulated gathers and scatters are tracked separately
from hardware gathers and scatters. Cost model heuristics for software
gathers/scatters is controlle with its own threshold.

As a temporal measure the new code is limitted to kick in for i16
data types only and for VFs > 2 and for cost < 100.