[VPlan][CfgMerger] Use merge PHI for dynamic peel count after peel loop. (#13227)
In the dynamic peel case, when emitting peel count checks after the peel
loop, use the merge PHI corresponding to the peel count, instead of using
the computed peel count directly, i.e:

```
; Old VPlan IR
  <peel-check>:
   i64 <computed-peel> = ...
   i1 %vp42484 = icmp eq i64 0 i64 <computed-peel>
   br i1 %vp42484, <after-peel>, <peel-bb>

    <after-peel>:
     i64 <peel-merge-phi> = phi-merge  [ i64 0, <peel-check> ],  [ i64 <peel-out>, <peel-bb> ]
     br BB25

    BB25:
     i64 %vp42836 = add i64 <computed-peel> i64 4
     i1 %vp34496 = icmp ugt i64 %vp42836 i64 %vp29734
     br i1 %vp34496, merge.blk19, BB1

; New VPlan IR
  <peel-check>:
   i64 <computed-peel> = ...
   i1 %vp42484 = icmp eq i64 0 i64 <computed-peel>
   br i1 %vp42484, <after-peel>, <peel-bb>

    <after-peel>:
     i64 <peel-merge-phi> = phi-merge  [ i64 0, <peel-check> ],  [ i64 <peel-out>, <peel-bb> ]
     br BB25

    BB25:
     i64 %vp42836 = add i64 <peel-merge-phi> i64 4
     i1 %vp34496 = icmp ugt i64 %vp42836 i64 %vp29734
     br i1 %vp34496, merge.blk19, BB1
```

This is to account for the fact that some paths may have
skipped the peel loop, in which case, the actual peel count will be zero
(despite the computed peel count *not* being zero)

This is part of a series of patches intended to support emitting a trip
count check to skip the peel loop. This patch should have no effect on
runtime behavior, as currently the only time we skip the peel loop is
when the computed peel count is zero.