[VPlan] Add option to skip dynamic peeling when trip count is less than specified threshold. (#13236)
Add an option `-vplan-skip-dynamic-peel-tc` (default: 0) which controls
the trip count threshold under which we skip dynamic peeling. When this
value is non-zero, we emit a runtime check before all other peel checks
to see if the loop TC is lesser than this value. If so, we skip directly
to the main loop without peeling.

JR: CMPLRLLVM-47569.