Introduce Intel_UnpredictableProfileLoader for PMU branch-miss profiles (#13155)
This pass reads simple IP-based profiles of branch-miss PMU events and
uses them to emit !unpredictable metadata. This can be thought of as
automatically adding __builtin_unpredictable() hints on branch
conditions based on PMU feedback.

On Linux, such a profile may be created with something like:

    perf record -b -c 10007 -e branch-misses:uppp ...
    llvm-profgen --leading-ip-only --perfdata perf.data ...

This profile file is completely separate from a traditional PGO or SPGO
profile, though it shares the same format.

Nothing generates or uses these profiles automatically yet. The
parameters early adopters are likely to find useful are:

 - `llvm-profgen --trim-cold-profile --profile-summary-cutoff-cold=<percentile>`
 - `--unpredictable-hints-min-weight=<count>`