[SYCL][CUDA] Map maxrregcount build option to a Cuda JIT option for use with ptxjit (#9139) (#13178)
This PR maps the "--maxrregcount" (from ptxas) compiler option string
token to a Cuda JIT option (`CUjit_option`) for use with
`ptxjitcompiler` to limit register pressure.

The idea is that for now, while not a complete solution, this can help
edge-cases where the use of too many registers not spilling above
certain threshold needs to be handled for the kernel to even be able to
launch.

It is related to a PR that handles this edge case by propagating the
back-end API error and throwing a targeted SYCL exception: #9106. It is
also one possible solution to the jit binary from issue #8414 and to one
of the issues found in the Blender project.

This string can currently be passed via the
`SYCL_PROGRAM_COMPILE_OPTIONS` environment variable when executing the
binary, i.e.:
```
SYCL_PROGRAM_COMPILE_OPTIONS="--maxrregcount=64" ./a.out
```

Ideally, compiler front-end to runtime propagation would make this
better rather than using the `env`, but the mapping to the appropriate
`CUjit_option` still needs to be handled similarly in the Cuda PI.

---

Still figuring out some appropriate tests for this change. It does fix
the kernel launching errors mentioned in the above issues.

--
cherry pick for codeplay 2023.2
JIRA: CMPLRLLVM-47823
Xmain: bf5749f89c9bf3c7f89751ca660567e9f081d85a (20230501_201948 pulldown)
Co-authored-by: Georgi Mirazchiyski <georgi.mirazchiyski@codeplay.com>