CMPLRLLVM-47181: Update libcxx lit test config for shared libs build (#13249)
This patch updates the config file for libcxx lit tests when the compiler
is being built for shared libraries. These lit tests overload the operator
'new' to modify a global variable. The compiler assumes that 'new' won't
modify neither will refer a global variable since it use the information
from C++ std. The changes adds the flag -fno-assume-sane-operator-new
to the libcxx lit tests to specify the compiler that overloaded new is not
C++ conformant.

This is the same patch as PR #13199, but for xmain-rel.

JIRA: CMPLRLLVM-47181