CMPLRLLVM-27323: 538.imagick IP Predicate Opt #23: Inlining report support (#13416)
* CMPLRLLVM-27323: 538.imagick PredicateOpt #23: Inline report support

This provides doOutlining() and replaceAllUsesWith() member functions
for the metadata inlining report and uses them in the community partial
inliner and cloning code (including 538.imagick IP predicate opt). A fix
for both inlining reports for the IP Predicate Opt multiversioning is
included. The LIT tests are updated, primarily to test for the output of
the metadata inlining report.

* Clang format

* Fix typos

* Fix suggested by Prasad

* Clang format