Introduce Intel_UnpredictableProfileLoader and Intel_AggressiveSpeculation (#13155 and #13145) (#13290)
This PR combines two separately reviewed xmain changes to avoid reintroducing/re-resolving merge conflicts.

Intel_UnpredictableProfileLoader reads simple IP-based profiles of branch-miss PMU events and
uses them to emit !unpredictable metadata. This can be thought of as
automatically adding __builtin_unpredictable() hints on branch
conditions based on PMU feedback.

On Linux, such a profile may be created with something like:

    perf record -b -c 10007 -e branch-misses:uppp ...
    llvm-profgen --leading-ip-only --perfdata perf.data ...

This profile file is completely separate from a traditional PGO or SPGO
profile, though it shares the same format.

Nothing generates or uses these profiles automatically yet. The
parameters early adopters are likely to find useful are:

 - `llvm-profgen --trim-cold-profile --profile-summary-cutoff-cold=<percentile>`
 - `--unpredictable-hints-min-weight=<count>`

Intel_AggressiveSpeculation can aggressively speculate then/else blocks such that control flow is eliminated but data flow is preserved via "select" instructions.

E.g.,:

    int v;
    if (__builtin_unpredictable(c)) {
      int r = a + 3;
      v = call(r);
    } else {
      v = call(7) + 10;
    }
    return v;

becomes:

    int v;

    int r = a + 3;          // speculative from "then"
    v = call(c ? r : 7);    // merged from both
    int v_false = v + 10;   // speculative from "else"

    return c ? v : v_false;

Because this results in the speculation of entire basic blocks, we're unlikely to be able to get profitability right based solely on static analysis. As a result, this transformation is considered to be profitable only when we have "[unpredictable](https://llvm.org/docs/LangRef.html#unpredictable-metadata)" metadata on the IR, which typically appears as a result of a programmer hint (see above) or upcoming PMU-based LLVM profile feedback. (See #13155.)

This initial implementation of the transformation can recognize and merge at most 1 non-speculatable function call somewhere in each of the then/else blocks. This is a simplification to support the first application of interest, and could be made more general in the future.

Toward Jiras: CMPLRLLVM-47539, CMPLRLLVM-47089, CMPLRLLVM-45473