[NFC] Update LIT testing for unified layout differences (#13449)
A number of tests are expecting a specific layout of the directories
used when performing compilation.  The INTEL_DEPLOY_UNIFIED_LAYOUT
setting when building the compiler will cause the compiler to expect a
different directory structure.

Update the testing to expect this build different possibility and update
the corresponding tests to enable proper testing in that environment.