[SYCLTransforms] Use BarrierRegionInfo to reduce load from special buffer (#13545)
This PR replaces old approach with BarrierRegionInfo which is more
efficient. If a cross-barrier value is used in multiple basic blocks
within a barrier region, it is only necessary to load the value from
special buffer once in region header.

The move of BarrierRegionInfo from ReduceCrossBarrierValues.cpp to
BarrierRegionInfo.h is a NFC change.

BarrierRegionInfo reduces shared-func-barrier-buffer-size.tst time from
794s to 338s in debug build. This PR further reduces test time to 22s by
simplifying the test workload. Jira: CMPLRLLVM-46603