Improve FPGA archive device unbundling with AOCO (#9572) (#13556)
AOCO specific objects can be added to an archive.  When this occurs, the archive was being completely ignored, causing issues with device compilation.  Take advantage of the additional `-excluded-targets` option from the `clang-offload-bundler` so we can process these archives for the needed device information, excluding potential unbundling of the target device in the presence of `aoco` binaries in that same object.

---------

Cherry-pick of intel/llvm#9572
Co-authored-by: Michael Toguchi <michael.d.toguchi@intel.com>