[PGO] Improve PGO icall promotion to convert additional cases. (#13574)
* [PGO] Improve PGO icall promotion to convert additional cases.

This extends the PGO indirect call promotion to better match the classic
compiler behavior for 511.povray. In icc, there were 2 targets converted
from the function All_CSG_Union_Intersections, but xmain was only
converting one of the targets due the heurstic that required a specific
percentage of the total calls. However, because this other call was
the only other target collected for that callsite, it should get
converted as well to avoid taking the indirect call path.

Also, improves the debug traces for the indirect promotion pass.

[CMPLRLLVM-48182]

* Fix comment typo