CMPLRLLVM-23157: Enable hoisting inside the Else branch after partial unswitching (#13579)
Originally, we didn't allowed more unswitching for a loop after applying
partial unsiwtching. One of the reasons is to make sure the level of the
conditions is preserved. This patch allows to keep applying unswitching
after partial unswitching for invariant predicates. It will be done only
for Else branch, since it is the part that will produce a result in the
same level.

CMPLRLLVM-23157