[MergeLoopExits][EarlyExitVec] Extend VPlan's loop massaging transforms for early-exit outer loops. (#13476)
Currently the massaging is restricted only to inner multi-exit
loops. This PR extends it to outer loops as well whenever
explicit early-exit loop vectorization is enabled. This allows
us to capture the condition-bit that triggers early exit and
represent that explicitly in VPlan IR, thereby laying the
foundation for upcoming transformEarlyExitLoops. This is the
second patch in the series for the early-exit loop vectorization
feature.

JIRA: CMPLRLLVM-47567
Change-Id: I251461fd88e70c3742835121721f6fa6cfb85137