CMPLRLLVM-48218: Fix the paramaters passed to hasEqualParentNode in HIROptPredicate (#13662)
The pramaters to the function hasEqualParentNode are HLNode, HLLoop,
unsigned and bool. The callsite was passing HLNode, HLLoop, bool and
unsigned. The unsigned parameter is used the predicate position during
partial unswitching for invariant predicate. Which means that even if
CheckPredPosOnly is true, the position will be incorrect when
PUCPredPos is higher than 1.