[Paropt] Expand master thread regions (#13411)
Instead of only specifically guarding contiguous sets of instructions
with non-thread-local side-effects, this change implements a new
approach to master thread guards which attempts to expand the guarded
regions as far as possible without requiring extra broadcasts of local
values. Expanding master thread regions reduces the number of master
thread guards needed to guard all of the non-thread-local side-effect
instructions in a target region, which simplifies the code and reduces
the number of redundant barriers added around the guarded regions.

To help analysis possible performance regressions, it's possible to disable
this new strategy and revert to the old one with the
-mllvm -vpo-paropt-master-thread-region-expansion=false option.

Jira: CMPLRLLVM-45504

This also fixes or helps to fix several reported issues with excessive
redundant barriers, including CMPLRLLVM-35930, CMPLRLLVM-43649, and
CMPLRLLVM-43607.