[VPlan] Add option to compute minimum profitable trip count for dynamic peeling. (#13418)
Add a switch `-vplan-calc-min-profitable-dyn-peel` (default: false)
which enables the computation of a minimum profitable trip count for
dynamic peel loops. This computation yields a threshold value which is
used for the run-time TC check prior to peeling, when
`-vplan-skip-dynamic-peel-tc` is not specified (see PR #13236).

This threshold is computed by solving an inequality of the form
```
[cost-of-loop-without-peel] < [cost-of-loop-with-peel]
```

in terms of the scalar trip count, obtaining a threshold value such that
the above inequality holds when
```
[scalar-tc] < [threshold] (i.e. the minimum profitable peel tc)
```

Additionally, the switch `-vplan-max-profitable-dyn-peel-multiplier`
(default: 150) was also added, to truncate large values. This works by
limiting the computed value to this switch value * VF * UF.

We additionally add a field to SingleLoopVectorizationScenario to hold
the computed minimum threshold, which is subsequently consumed by the
CFG merger to emit the TC check.

JR: CMPLRLLVM-47569.