[VPlan][VPSE] Check higher dims are equal before computing diff SCEVs. (#13722)
* [VPlan][VPSE] Check higher dims are equal before computing diff SCEVs.

When computing the diff between two address SCEVs, use the original ref
to check that the higher dimensions (> 1) are equal, before computing a
valid difference between them. If they are not, return `nullptr`.

This works around an issue in computing address SCEVs where we ignore
the upper dimensions when computing the adjusted invariant base for the
recurrence. This issue meant that multi-dimensional memrefs with
differing upper dimensions (e.g. `A[K][J-1][I]` and `A[K][J+1][I]`)
would sometimes incorrectly be computed as congruent. In practice, this
would sometimes cause VPlan to favor peeling the wrong memref (e.g. a
load over a store), due to incorrectly believing that peeling that
memref would align other congruent memrefs.

A more long-term fix would be to correctly add the upper dimensions when
computing the invariant base. That was not attempted here, but a TODO
was left to remark that we may wish to continue that work in the future.

JR: CMPLRLLVM-48366.